re #32 adding period update class in EpicsClient
