+++ src/EpicsClient/EpicsClientPeriodic.cpp
@@ -1,0 +1,14 @@
+#include "EpicsClientPeriodic.h"
+
+namespace BrightnESS {
+namespace ForwardEpicsToKafka {
+namespace EpicsClient {
+
+EpicsClientPeriodic::EpicsClientPeriodic(Stream *stream,
+                                         std::string epics_channel_provider_type,
+                                         std::string channel_name) {
+
+}
+}
+}
+}
