+++ src/EpicsClient/EpicsClientPeriodic.h
@@ -1,0 +1,22 @@
+#pragma once
+#include <pv/pvaClient.h>
+
+namespace BrightnESS {
+namespace ForwardEpicsToKafka {
+namespace EpicsClient {
+class Stream;
+class ForwarderInfo;
+
+class EpicsClientPeriodic {
+public:
+  std::shared_ptr<ForwarderInfo> finfo;
+  Stream *stream = nullptr;
+  EpicsClientPeriodic(Stream *stream,
+                      std::string epics_channel_provider_type,
+                      std::string channel_name);
+private:
+//  epics::pvaClient::PvaClientGet;
+};
+}
+}
+}
